split out startPointItem class to its own file
