+++ Lib/defconQt/glyphView.py
@@ -24,9 +24,9 @@ from defconQt.layerActionsDialog import LayerActionsDialog
 from defconQt.offCurvePointItem import OffCurvePointItem
 from defconQt.onCurvePointItem import OnCurvePointItem
 from defconQt.pixmapItem import PixmapItem
+from defconQt.startPointItem import StartPointItem
 from defconQt.vGuidelinesTextItem import VGuidelinesTextItem
 
-startItemDist = 10
 backgroundColor = Qt.white
 bluesColor = QColor.fromRgbF(.5, .7, 1, .3)
 fillColor = QColor(200, 200, 200, 120)  # QColor.fromRgbF(0, 0, 0, .4)
@@ -35,47 +35,6 @@ componentFillColor = QColor.fromRgbF(
 metricsColor = QColor(70, 70, 70)
 
 bluesAttrs = ["postscriptBlueValues", "postscriptOtherBlues"]
-
-class StartPointItem(QGraphicsPathItem):
-
-    def __init__(self, x, y, angle, scale=1, parent=None):
-        super(StartPointItem, self).__init__(parent)
-        self._angle = 360 - angle
-
-        self.setPointPath(scale)
-        self.setPos(x, y)
-        self.setZValue(-996)
-
-    def setPointPath(self, scale=None):
-        if scale is None:
-            scene = self.scene()
-            if scene is not None:
-                scale = scene.getViewScale()
-            else:
-                scale = 1
-        if scale > 1.30:
-            scale = 1.30
-        elif scale < .6:
-            scale = .6
-        self.prepareGeometryChange()
-        dist = startItemDist / scale
-        path = QPainterPath()
-        line = QLineF(0, 0, 0 + dist, 0)
-        line2 = QLineF(line)
-        line.setAngle(self._angle - 90)
-        path.lineTo(line.x2(), line.y2())
-        line2.setAngle(self._angle)
-        line2.translate(line.p2() - line.p1())
-        path.lineTo(line2.x2(), line2.y2())
-        line.setP1(line2.p2())
-        line.setAngle(line.angle() - 27.5)
-        line.setLength(2 * dist / 5)
-        line2.setLength(line2.length() + .5)
-        path.moveTo(line.x2(), line.y2())
-        path.lineTo(line2.x2(), line2.y2())
-        line.setAngle(line.angle() + 55)
-        path.lineTo(line.x2(), line.y2())
-        self.setPath(path)
 
 class GlyphView(QGraphicsView):
